applied suggestions from pr comments

+++ models/inputs/addition.go
@@ -4,5 +4,5 @@ type Addition struct {
 	NodeURL  string `json:"node_url"`
 	Standard string `json:"standard"`
 	Owner    string `json:"owner"`
-	Number   int64  `json:"number"`
+	Number   uint   `json:"number"`
 }
